nsqadmin: refactor /graphite
